Add files via upload
In Chile, the Multidimensional Poverty Index (MPI) has been used since 2015 to characterize households. However, even though the multidimensional poverty approach diverges from the reductionism determined by the measurement of poverty by income, it is necessary to investigate the way in which the structures of deprivation are configured in the different socioeconomic strata. This study aims to elucidate the correlation between the level of income and the affliction of different deprivations in the households interviewed in the CASEN 2017 survey, as well as the usefulness of the MPI in expressing their level of vulnerability. To this end, a typological study was conducted on the deficiencies of multidimensional poor households in each socioeconomic strata, as well as the correlation between the MPI and income level. The results demonstrated that, although there is a greater prevalence of multidimensional poverty in lower-income households, this is not expressed in higher intensity of the MPI, with the majority of the households being near the threshold. In contrast, as income levels increase, there is a shift in deficiencies from the dimensions of “Education” and “Housing and Environment” to those of “Health” and “Employment and social security”