Add files via upload
The concept of poverty has transited from a purely economic approach to a multidimensional one, which allows the application of systemic analysis to the phenomenon. In this context, it is not only necessary to analytically understand the prevalence of deprivations of different natures, but it is also relevant to study the relationships and emerging properties between the internal components of the phenomenon. This article seeks to systematize the relational methodological proposal for the study of multidimensional poverty and to update the analysis of the internal structure of multidimensional poverty in Chile, using the CASEN Survey 2017. In addition, this work inquires into the effect it produces on the analysis of the categorization of households into different types. The main conclusions of the study are: (i) multidimensional poverty presents a complex behavior, and its study allows observing new characteristics of deprivations, unnoticed from the analytical perspective; (ii) the deprivations "equal treatment" and "security", of the dimension "networks and social cohesion", are the ones that have the greatest structural importance in the configuration of the phenomenon of multidimensional poverty; iii) the application of the property of decomposability by household type makes it possible to identify unique characteristics for each subgroup studied; and iv) the development of tools for the systemic analysis of poverty implies a rethinking of the current way of designing and implementing public policies.